+++ programs/devolt/src/lib.rs
@@ -17,12 +17,12 @@ pub mod devolt {
         id: String,
         latitude: Option<f64>,
         longitude: Option<f64>,
-        capacity: Option<u64>,
-        available: u64,
+        capacity: Option<f64>,
+        available: f64,
     ) -> Result<()> {
         msg!("Reporting battery at station with ID: {}\nNew availability: {}", id, available);
 
-        require!(id != "" || available != 0, errors::DevoltError::InvalidInput);
+        require!(id != "" || available != 0.0, errors::DevoltError::InvalidInput);
 
         instructions::battery_report::battery_report(ctx, id, latitude, longitude, capacity, available)
     }
